% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpride2.R
\name{pride_1}
\alias{pride_1}
\title{Draw a bar graph about who will be given the time of day around here}
\usage{
pride_1()
}
\description{
Draw a bar graph about who will be given the time of day around here
}
